tests: structure
