+++ pom.xml
@@ -116,6 +116,11 @@
       <artifactId>jacoco-maven-plugin</artifactId>
       <version>0.8.8</version>
     </dependency>
+    <dependency>
+      <groupId>com.github.stefanbirkner</groupId>
+      <artifactId>system-lambda</artifactId>
+      <version>1.2.1</version>
+    </dependency>
   </dependencies>
 
   <build>
