+++ projet-louis_mondaini_nunez.iml
@@ -56,5 +56,6 @@
     <orderEntry type="library" name="Maven: org.ow2.asm:asm-analysis:9.2" level="project" />
     <orderEntry type="library" name="Maven: org.ow2.asm:asm-tree:9.2" level="project" />
     <orderEntry type="library" name="Maven: org.jacoco:org.jacoco.report:0.8.8" level="project" />
+    <orderEntry type="library" name="Maven: com.github.stefanbirkner:system-lambda:1.2.1" level="project" />
   </component>
 </module>
