--- site/config.yaml
@@ -1,3 +1,0 @@
-titre: titre du site
-description: lorem ipsum
-domain: example.com
