--- site/index.md
@@ -1,7 +1,0 @@
-titre: Mon premier article
-auteur: John Doe
-date: 2022-04-28
----
-# Mon premier article
-## Mon sous-titre
-Le contenu de mon article.
