--- site/template/layout.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>{{ site.titre }} | {{ page.titre }}</title>
-</head>
-<body>
-{{> menu }}
-{{ content }}
-</body>
-</html>
