--- site/template/menu.html
@@ -1,3 +1,0 @@
-<ul>
-    <a href="index.html">Index</a>
-</ul>
